data migration for billAddress,users,payments table

+++ src/main/resources/db/migration/V1.0.1__Alter_Users_username.sql
@@ -1,4 +1,3 @@
-
 ALTER TABLE users ADD COLUMN username varchar(255) NOT NULL;
-ALTER TABLE users ADD COLUMN email  varchar(255) NOT NULL ;
-ALTER TABLE users ADD COLUMN phone  numeric (20) NOT NULL;
+ALTER TABLE users ADD COLUMN email varchar(255) NOT NULL ;
+ALTER TABLE users ADD COLUMN phone integer NOT NULL;
