+++ src/main/resources/db/migration/V1.1.1__Add_Payments.sql
@@ -1,7 +1,7 @@
 CREATE SEQUENCE payments_id_seq;
 create table payments (
    id bigint not null DEFAULT NEXTVAL('payments_id_seq'),
-   card_number numeric (16) NOT NULL,
+   card_number integer NOT NULL,
    card_type varchar (25) NOT NULL,
    holder_first_name varchar(255) not NULL,
    holder_last_name varchar(255) not NULL,
