+++ src/main/resources/db/migration/V1.1.2__Add_BillAddress.sql
@@ -3,10 +3,11 @@ create table bill_address (
    id bigint not null DEFAULT NEXTVAL('bills_id_seq'),
    address1 varchar (255) NOT NULL,
    address2 varchar (255),
-   zip_code numeric (10) NOT NULL,
+   zip_code Integer NOT NULL,
    city varchar(255) NOT NULL,
    state varchar(255) NOT NULL,
    country varchar(255) NOT NULL,
-   phone_number numeric (20) NOT NULL,
-   primary key (id));
+   phone_number integer NOT NULL,
+   primary key (id)
+   );
 ALTER SEQUENCE bills_id_seq OWNED BY bill_address.id;
